WIP: refactoring style InUseSizeLocker modal
