--- src/modals/InUseModalSizeLocker.jsx
@@ -1,39 +1,0 @@
-import "../styles/styles.css";
-import React, { useState, useEffect } from "react";
-
-export const InUseModalSizeLocker = () => {
-    const [isModalOpen, setModalOpen] = useState(false);
-
-    const openModal = () => setModalOpen(true);
-    const closeModal = () => setModalOpen(false);
-
-    return (
-        <div id="modal-container" className="fixed inset-0 flex items-center justify-center sm:px-6 lg:px-20">
-            <div id="modal-content" className="bg-[#BEC7FF] p-8 rounded shadow-md w-[40%] text-center">
-                <p className="py-5 text-6xl">❌ </p>
-                <h1 className="text-base sm:text-xl font-semibold mb-6 whitespace-nowrap overflow-hidden text-ellipsis">
-                    Los lockers de ese tamaño están ocupados
-                </h1>
-                <div id="modal-buttons-inUseLocker" className="flex flex-col sm:flex-row justify-center gap-2">
-                    <button id="button-choose-size" onClick={openModal} className="w-[36%] bg-[#FFFFFF] text-black py-2 px-4 rounded border hover:opacity-90 transition font-bree text-center text-sm sm:text-base cursor-pointer">
-                        ESCOGER TAMAÑO
-                    </button>
-                    <button id="botton-exit" onClick={closeModal} className="w-[36%] bg-[#00174B] text-white py-2 px-4 rounded border hover:opacity-90 transition font-bree text-center text-sm sm:text-base cursor-pointer">
-                        SALIR
-                    </button>
-                </div>
-            </div>
-        </div>
-    );
-};
-
-
-
-
-
-
-
-
-
-
-
